🔥 Cleanup + migrate trident svm
