+++ Cargo.toml
@@ -1,36 +1,22 @@
 [workspace]
-members = ["crates/cli", "crates/client", "crates/fuzz"]
+members = ["crates/cli", "crates/client", "crates/config", "crates/fuzz"]
 exclude = ["examples/"]
 resolver = "1"
 
 
 [workspace.dependencies]
-# ANCHOR
-anchor-lang-idl-spec = "0.1.0"
-
-# HONGGFUZZ
-honggfuzz = { version = "0.5.55" }
-
-#AFL
-afl = { version = "0.15.10" }
+# Solana
+solana-sdk = "~2.0"
+spl-token = { version = "5", features = ["no-entrypoint"] }
+solana-vote-program = "~2.0"
+solana-stake-program = "~2.0"
+solana-svm = "~2.0"
 
 # ARBITRARY
 arbitrary = { version = "1.3.0", features = ["derive"] }
 
-# SOLANA
-solana-sdk = "=2.1.0"
-solana-banks-client = "=2.1.0"
-solana-program-runtime = "=2.1.0"
-solana-vote-program = "=2.1.0"
-solana-stake-program = "=2.1.0"
-solana-system-program = "=2.1.0"
-solana-compute-budget = "=2.1.0"
-solana-svm = "=2.1.0"
-solana-bpf-loader-program = "=2.1.0"
-solana-config-program = "=2.1.0"
-solana-compute-budget-program = "=2.1.0"
-solana-address-lookup-table-program = "=2.1.0"
-solana-loader-v4-program = "=2.1.0"
-solana-logger = "=2.1.0"
-spl-associated-token-account = { version = "2", features = ["no-entrypoint"] }
-spl-token = { version = "5", features = ["no-entrypoint"] }
+# ARBITRARY
+honggfuzz = { version = "0.5.55" }
+
+#AFL
+afl = { version = "0.15.10" }
