+++ crates/client/src/commander/honggfuzz.rs
@@ -4,6 +4,7 @@ use std::process;
 use std::{os::unix::process::CommandExt, process::Stdio};
 use tokio::process::Command;
 
+use trident_config::honggfuzz::EnvVariable;
 use trident_config::Config;
 
 use crate::constants::*;
@@ -26,11 +27,11 @@ impl Commander {
 
         let mut fuzz_args = config.get_honggfuzz_args(hfuzz_run_args);
 
-        let cargo_target_dir =
-            std::env::var("CARGO_TARGET_DIR").unwrap_or_else(|_| config.get_honggfuzz_target_dir());
+        let cargo_target_dir = std::env::var("CARGO_TARGET_DIR")
+            .unwrap_or_else(|_| config.get_env_arg(&EnvVariable::CargoTargetDir));
 
-        let hfuzz_workspace =
-            std::env::var("HFUZZ_WORKSPACE").unwrap_or_else(|_| config.get_honggfuzz_workspace());
+        let hfuzz_workspace = std::env::var("HFUZZ_WORKSPACE")
+            .unwrap_or_else(|_| config.get_env_arg(&EnvVariable::HfuzzWorkspace));
 
         let (crash_dir, ext) =
             get_crash_dir_and_ext(&self.root, &target, &fuzz_args, &hfuzz_workspace);
@@ -90,10 +91,10 @@ impl Commander {
 
         let hfuzz_run_args = std::env::var("HFUZZ_RUN_ARGS").unwrap_or_default();
 
-        let cargo_target_dir =
-            std::env::var("CARGO_TARGET_DIR").unwrap_or_else(|_| config.get_honggfuzz_target_dir());
-        let hfuzz_workspace =
-            std::env::var("HFUZZ_WORKSPACE").unwrap_or_else(|_| config.get_honggfuzz_workspace());
+        let cargo_target_dir = std::env::var("CARGO_TARGET_DIR")
+            .unwrap_or_else(|_| config.get_env_arg(&EnvVariable::CargoTargetDir));
+        let hfuzz_workspace = std::env::var("HFUZZ_WORKSPACE")
+            .unwrap_or_else(|_| config.get_env_arg(&EnvVariable::HfuzzWorkspace));
 
         let mut fuzz_args = config.get_honggfuzz_args(hfuzz_run_args);
 
@@ -151,8 +152,8 @@ impl Commander {
             throw!(Error::CrashFileNotFound);
         }
 
-        let cargo_target_dir =
-            std::env::var("CARGO_TARGET_DIR").unwrap_or_else(|_| config.get_honggfuzz_target_dir());
+        let cargo_target_dir = std::env::var("CARGO_TARGET_DIR")
+            .unwrap_or_else(|_| config.get_env_arg(&EnvVariable::CargoTargetDir));
 
         let mut rustflags = std::env::var("RUSTFLAGS").unwrap_or_default();
