+++ crates/client/src/idl_loader.rs
@@ -37,6 +37,7 @@ pub fn load_idls(dir_path: PathBuf) -> Result<Vec<Idl>, Box<dyn Error>> {
                     // Continue to the next file on failure
                 }
             }
+            // }
         }
     }
