--- crates/client/src/source_code_generators/anchor_29/mod.rs
@@ -1,16 +1,0 @@
-mod custom_types;
-mod fuzz_accounts;
-pub mod fuzz_instructions_generator;
-mod idl_type_to_syn;
-mod instruction_inputs;
-mod instruction_ixops;
-mod instruction_variants;
-pub mod test_fuzz_generator;
-
-use custom_types::*;
-use fuzz_accounts::*;
-use idl_type_to_syn::*;
-use instruction_inputs::*;
-use instruction_ixops::*;
-use instruction_variants::*;
-pub(crate) mod types;
