+++ crates/client/src/template/Cargo_fuzz.toml.tmpl
@@ -1,3 +1,5 @@
+[workspace]
+
 [package]
 name = "fuzz_tests"
 version = "0.1.0"
@@ -7,3 +9,4 @@ edition = "2021"
 [dependencies]
 arbitrary = "1.3.0"
 borsh = "1.5.3"
+solana-sdk = "~2.0"
