+++ crates/config/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod afl;
-pub mod argument;
 pub mod constants;
 pub mod fuzz;
 pub mod honggfuzz;
@@ -9,12 +8,14 @@ use constants::*;
 use fuzz::*;
 use honggfuzz::*;
 
-pub mod utils;
-
+use anyhow::Context;
+use fehler::throw;
 use serde::Deserialize;
-use std::{fs, io};
+use std::{
+    env, fs, io,
+    path::{Path, PathBuf},
+};
 use thiserror::Error;
-use utils::{discover_root, resolve_path};
 
 #[derive(Error, Debug)]
 pub enum Error {
@@ -30,9 +31,29 @@ pub enum Error {
 
 #[derive(Debug, Deserialize, Clone)]
 pub struct Config {
-    pub honggfuzz: Option<HonggFuzz>,
-    pub afl: Option<Afl>,
-    pub fuzz: Option<Fuzz>,
+    pub honggfuzz: HonggFuzz,
+    pub afl: Afl,
+    pub fuzz: Fuzz,
+}
+
+#[derive(Default, Debug, Deserialize, Clone)]
+struct _Config {
+    #[serde(default)]
+    pub honggfuzz: Option<_HonggFuzz>,
+    #[serde(default)]
+    pub afl: Option<_Afl>,
+    #[serde(default)]
+    pub fuzz: Option<_Fuzz>,
+}
+
+impl From<_Config> for Config {
+    fn from(_c: _Config) -> Self {
+        Self {
+            honggfuzz: _c.honggfuzz.unwrap_or_default().into(),
+            afl: _c.afl.unwrap_or_default().into(),
+            fuzz: _c.fuzz.unwrap_or_default().into(),
+        }
+    }
 }
 
 impl Default for Config {
@@ -46,130 +67,105 @@ impl Config {
         let root = discover_root().expect("failed to find the root folder");
         let s = fs::read_to_string(root.join(TRIDENT_TOML).as_path())
             .expect("failed to read the Trident config file");
-        let _config: Config = toml::from_str(&s).expect("failed to parse the Trident config file");
-        _config
+        let _config: _Config = toml::from_str(&s).expect("failed to parse the Trident config file");
+        _config.into()
     }
 
-    // -*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*
-    // honggfuzz
     pub fn get_honggfuzz_args(&self, cli_input: String) -> String {
-        if let Some(honggfuzz) = &self.honggfuzz {
-            let mut args = honggfuzz.get_collect_fuzz_args();
-            args.push(cli_input);
-            args.join(" ")
-        } else {
-            String::default()
-        }
+        let mut args = self.honggfuzz.get_collect_fuzz_args();
+        args.push(cli_input);
+        args.join(" ")
     }
-    pub fn get_honggfuzz_workspace(&self) -> String {
-        let path = self
-            .honggfuzz
-            .as_ref()
-            .map(|honggfuzz| honggfuzz.get_hfuzz_workspace().value)
-            .unwrap_or_else(|| HFUZZ_WORKSPACE_DEFAULT_HFUZZ.to_string());
-        let full_path = resolve_path(&path);
-        full_path.to_str().unwrap().to_string()
+    pub fn get_env_arg(&self, key: &EnvVariable) -> String {
+        let expect = format!("{:#?} not found", key);
+        self.honggfuzz.get_env_variable(key).expect(&expect)
     }
-    pub fn get_honggfuzz_target_dir(&self) -> String {
-        let path = self
-            .honggfuzz
-            .as_ref()
-            .map(|honggfuzz| honggfuzz.get_cargo_target_dir().value)
-            .unwrap_or_else(|| CARGO_TARGET_DIR_DEFAULT_HFUZZ.to_string());
-        let full_path = resolve_path(&path);
-        full_path.to_str().unwrap().to_string()
-    }
-    // -*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*
-    // afl
-    pub fn get_afl_target_dir(&self) -> String {
-        let path = self
-            .afl
-            .as_ref()
-            .map(|afl| afl.get_cargo_target_dir().value.unwrap())
-            .unwrap_or_else(|| CARGO_TARGET_DIR_DEFAULT_AFL.to_string());
-        let full_path = resolve_path(&path);
-        full_path.to_str().unwrap().to_string()
-    }
-    pub fn get_afl_target_path(&self, target: &str) -> String {
-        let mut afl_target_dir = self.get_afl_target_dir();
-        afl_target_dir.push_str("/debug/");
-        afl_target_dir.push_str(target);
-        afl_target_dir
-    }
-    pub fn get_afl_workspace_in(&self) -> String {
-        let path = self
-            .afl
-            .as_ref()
-            .map(|afl| afl.get_workspace_in().value.unwrap())
-            .unwrap_or_else(|| AFL_WORKSPACE_DEFAULT_IN.to_string());
-        let full_path = resolve_path(&path);
-        full_path.to_str().unwrap().to_string()
-    }
-    pub fn get_afl_workspace_out(&self) -> String {
-        let path = self
-            .afl
-            .as_ref()
-            .map(|afl| afl.get_workspace_out().value.unwrap())
-            .unwrap_or_else(|| AFL_WORKSPACE_DEFAULT_OUT.to_string());
-        let full_path = resolve_path(&path);
-        full_path.to_str().unwrap().to_string()
-    }
+
     pub fn get_afl_build_args(&self) -> Vec<String> {
-        self.afl
-            .as_ref()
-            .map(|afl| afl.get_collect_build_args())
-            .unwrap_or_default()
+        self.afl.get_collect_build_args()
     }
     pub fn get_afl_fuzz_args(&self) -> Vec<String> {
+        self.afl.get_collect_fuzz_args()
+    }
+    pub fn get_afl_cargo_build_dir(&self) -> String {
         self.afl
-            .as_ref()
-            .map(|afl| afl.get_collect_fuzz_args())
-            .unwrap_or_default()
+            .get_cargo_build_dir()
+            .expect("AFL Cargo Target Dir argument not available")
+            .val
+            .clone()
+            .expect("AFL Cargo Target Dir value not available")
+    }
+    pub fn get_afl_target_path(&self) -> String {
+        let afl_arg = self
+            .afl
+            .get_cargo_build_dir()
+            .expect("AFL Cargo Target Dir argument not available");
+
+        let mut target_path = afl_arg
+            .val
+            .clone()
+            .expect("AFL Cargo Target Dir value not available");
+
+        target_path.push_str("/debug/");
+        target_path
+    }
+    pub fn get_afl_workspace_in(&self) -> String {
+        let afl_arg = self
+            .afl
+            .get_workspace_in()
+            .expect("AFL Workspace in value argument available");
+
+        afl_arg
+            .val
+            .clone()
+            .expect("AFL Workspace in value not available")
+    }
+    pub fn get_initial_seed(&self) -> &[AflSeed] {
+        &self.afl.seeds
     }
 
-    pub fn get_initial_seed(&self) -> Vec<AflSeed> {
-        self.afl
-            .as_ref()
-            .map(|afl| afl.get_seeds())
-            .unwrap_or_else(|| vec![AflSeed::default()])
-    }
-    // -*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*
-    // fuzz
     pub fn get_fuzzing_with_stats(&self) -> bool {
-        self.fuzz
-            .as_ref()
-            .map(|fuzz| fuzz.get_fuzzing_with_stats())
-            .unwrap_or_default()
+        self.fuzz.get_fuzzing_with_stats()
     }
     pub fn get_allow_duplicate_txs(&self) -> bool {
-        self.fuzz
-            .as_ref()
-            .map(|fuzz| fuzz.get_allow_duplicate_txs())
-            .unwrap_or_default()
+        self.fuzz.get_allow_duplicate_txs()
+    }
+}
+
+/// Tries to find the root directory with the `Anchor.toml` file.
+/// Throws an error when there is no directory with the `Anchor.toml` file
+pub fn discover_root() -> Result<PathBuf, Error> {
+    let current_dir = env::current_dir()?;
+    let mut dir = Some(current_dir.as_path());
+    while let Some(cwd) = dir {
+        for file in std::fs::read_dir(cwd)
+            .with_context(|| format!("Error reading the directory with path: {}", cwd.display()))?
+        {
+            let path = file
+                .with_context(|| {
+                    format!("Error reading the directory with path: {}", cwd.display())
+                })?
+                .path();
+            if let Some(filename) = path.file_name() {
+                if filename.to_str() == Some(ANCHOR_TOML) {
+                    return Ok(PathBuf::from(cwd));
+                }
+            }
+        }
+        dir = cwd.parent();
+    }
+    throw!(Error::BadWorkspace)
+}
+
+pub fn find_full_path(input: &str) -> Result<PathBuf, std::io::Error> {
+    let path = Path::new(input);
+
+    // If the path is absolute, return it directly
+    if path.is_absolute() {
+        return Ok(path.to_path_buf());
     }
 
-    pub fn programs(&self) -> Vec<FuzzProgram> {
-        self.fuzz
-            .as_ref()
-            .map(|fuzz| {
-                if let Some(programs) = &fuzz.programs {
-                    programs.iter().map(FuzzProgram::from).collect()
-                } else {
-                    Vec::default()
-                }
-            })
-            .unwrap_or_default()
-    }
-    pub fn accounts(&self) -> Vec<FuzzAccount> {
-        self.fuzz
-            .as_ref()
-            .map(|fuzz| {
-                if let Some(accounts) = &fuzz.accounts {
-                    accounts.iter().map(FuzzAccount::from).collect()
-                } else {
-                    Vec::default()
-                }
-            })
-            .unwrap_or_default()
-    }
+    let root = discover_root().expect("Failed to obtain Anchor root");
+
+    Ok(root.join(path))
 }
