+++ crates/fuzz/src/fuzz_client.rs
@@ -1,18 +1,18 @@
 #![allow(dead_code)]
 
+use trident_config::Config;
 use crate::error::*;
-use solana_program_runtime::invoke_context::BuiltinFunctionWithContext;
 use solana_sdk::account::AccountSharedData;
-use solana_sdk::account_info::AccountInfo;
-use solana_sdk::entrypoint::ProgramResult;
 use solana_sdk::hash::Hash;
 use solana_sdk::instruction::Instruction;
 use solana_sdk::pubkey::Pubkey;
 use solana_sdk::signature::Keypair;
 use solana_sdk::sysvar::Sysvar;
+use trident_svm::utils::ProgramEntrypoint;
 
 /// A trait providing methods to read and write (manipulate) accounts
 pub trait FuzzClient {
+    fn new_client(programs: &[ProgramEntrypoint], config: &Config) -> Self;
     /// Get the cluster rent
     fn get_sysvar<T: Sysvar>(&mut self) -> T;
 
@@ -38,53 +38,11 @@ pub trait FuzzClient {
     fn get_last_blockhash(&self) -> Hash;
 
     /// Send a transaction and return until the transaction has been finalized or rejected.
-    fn process_instruction(&mut self, _instruction: Instruction) -> Result<(), FuzzClientError>;
+    fn process_instructions(
+        &mut self,
+        _instructions: &[Instruction],
+    ) -> Result<(), FuzzClientError>;
 
     // Clear Temp account created during fuzzing iteration
     fn clear_accounts(&mut self);
-}
-
-pub type ProgramEntry = for<'info> fn(
-    program_id: &Pubkey,
-    accounts: &'info [AccountInfo<'info>],
-    instruction_data: &[u8],
-) -> ProgramResult;
-
-/// Converts Anchor 0.29.0 and higher entrypoint into the runtime's entrypoint style
-///
-/// Starting Anchor 0.29.0 the accounts are passed by reference https://github.com/coral-xyz/anchor/pull/2656
-/// and the lifetime requirements are `accounts: &'a [AccountInfo<'a>]` instead of `accounts: &'a [AccountInfo<'b>]`.
-/// The new requirements require the slice of AccountInfos and the contained Accounts to have the same lifetime but
-/// the previous version is more general. The compiler implies that `'b` must live at least as long as `'a` or longer.
-///
-/// The transaction data is serialized and again deserialized to the `&[AccountInfo<_>]` slice just before invoking
-/// the entry point and the modified account data is copied to the original accounts just after the the entry point.
-/// After that the `&[AccountInfo<_>]` slice goes out of scope entirely and therefore `'a` == `'b`. So it _SHOULD_ be
-/// safe to do this conversion in this testing scenario.
-///
-/// Do not use this conversion in any on-chain programs!
-#[macro_export]
-macro_rules! convert_entry {
-    ($entry:expr) => {
-        unsafe { core::mem::transmute::<ProgramEntry, ProcessInstruction>($entry) }
-    };
-}
-
-pub struct FuzzingProgram {
-    pub program_name: String,
-    pub program_id: Pubkey,
-    pub entry: Option<BuiltinFunctionWithContext>,
-}
-impl FuzzingProgram {
-    pub fn new(
-        program_name: &str,
-        program_id: &Pubkey,
-        entry_fn: Option<BuiltinFunctionWithContext>,
-    ) -> FuzzingProgram {
-        Self {
-            program_name: program_name.to_string(),
-            program_id: *program_id,
-            entry: entry_fn,
-        }
-    }
 }
