+++ crates/fuzz/src/lib.rs
@@ -3,10 +3,9 @@ pub mod error;
 pub mod fuzz_data;
 pub mod fuzz_stats;
 pub mod snapshot;
-pub mod trident_svm;
 pub type AccountId = u8;
-pub mod config;
 pub mod fuzz_client;
+pub mod fuzz_client_impl;
 pub mod fuzz_test_executor;
 pub mod fuzz_trident;
 pub mod instructions_sequence;
@@ -39,7 +38,6 @@ pub mod fuzzing {
     pub use trident_derive_fuzz_test_executor::FuzzTestExecutor;
 
     /// trident macros
-    pub use super::convert_entry;
     pub use super::fuzz_trident;
     pub use super::middle_sequence;
     pub use super::post_sequence;
@@ -47,12 +45,10 @@ pub mod fuzzing {
     pub use super::show_account;
     pub use super::*;
 
-    pub use super::fuzz_client::FuzzingProgram;
-    pub use super::fuzz_client::ProgramEntry;
-
     /// trident methods
     pub use super::accounts_storage::*;
-    pub use super::config::Config;
+    pub use trident_config::Config;
+
     pub use super::error::*;
     pub use super::fuzz_client::FuzzClient;
     pub use super::fuzz_data::build_ix_fuzz_data;
@@ -63,10 +59,11 @@ pub mod fuzzing {
     pub use super::snapshot::Snapshot;
     pub use super::snapshot::SnapshotAccount;
     pub use super::transaction_executor::TransactionExecutor;
-    pub use super::trident_svm::svm::TridentSVM;
 
     pub use std::cell::RefCell;
     pub use std::collections::HashMap;
+    pub use trident_svm::trident_svm::TridentSVM;
+    pub use trident_svm::utils::ProgramEntrypoint;
 
     pub use super::accounts_storage::KeypairStore;
     pub use super::accounts_storage::PdaStore;
