--- crates/fuzz/src/trident_svm/log.rs
@@ -1,8 +1,0 @@
-pub(crate) fn setup_solana_logging() {
-    #[rustfmt::skip]
-    solana_logger::setup_with_default(
-        "solana_rbpf::vm=debug,\
-            solana_runtime::message_processor=debug,\
-            solana_runtime::system_instruction_processor=trace",
-    );
-}
