--- crates/fuzz/src/trident_svm/mod.rs
@@ -1,5 +1,0 @@
-pub mod accounts_db;
-pub mod error;
-pub mod log;
-pub mod native;
-pub mod svm;
