+++ examples/hello_world/programs/hello_world/Cargo.toml
@@ -17,4 +17,4 @@ no-log-ix-name = []
 idl-build = ["anchor-lang/idl-build"]
 
 [dependencies]
-anchor-lang = { git = "https://github.com/coral-xyz/anchor", branch = "master" }
+anchor-lang = "0.30.1"
